add pkl2vtk for converting the approx 4x smaller rollout_xxx.pkl files to a series of vtk files for visualization
